kernel: improve create session code
Create session takes hlist lock to prevent creation of another sessions
with the same ip address. But the code needs a lot of time to fully
execute and creates lock contention and prevents traffic to pass during
new session creation. Improve it to do the minimal things need for the
session under the lock and move population of informational fields out
of the critical section.

Signed-off-by: Vadim Fedorenko <vvfedorenko@github.com>